Clean up modules, imports, ensure private modules aren't exported, run `black`
