+++ python/packages/sdk/sdk/__init__.py
@@ -3,8 +3,7 @@ from __future__ import annotations
 from typing_extensions import Final
 from typing import List
 
-from .sdk import ServerlessSdk
-from .trace_span import TraceSpan
+from .sdk.base import ServerlessSdk
 
 
 # public exports
