+++ python/packages/sdk/sdk/sdk/base.py
@@ -5,9 +5,8 @@ from typing import List, Optional
 
 from typing_extensions import Final
 
-from ..base import Nanoseconds, SLS_ORG_ID, __version__, __name__
+from ..base import Nanoseconds, SLS_ORG_ID, __version__
 from ..span.trace import TraceSpan
-from ..span.tags import Tags
 
 
 __all__: Final[List[str]] = [
@@ -19,20 +18,19 @@ class ServerlessSdk:
     name: Final[str] = __name__
     version: Final[str] = __version__
 
-    trace_spans: Final = ...
+    traceSpans: Final = ...
     instrumentation: Final = ...
 
-    org_id: Optional[str] = None
+    orgId: Optional[str] = None
 
     def _initialize(self, org_id: Optional[str] = None):
-        self.org_id = environ.get(SLS_ORG_ID, default=org_id)
+        self.orgId = environ.get(SLS_ORG_ID, default=org_id)
 
-    def create_trace_span(
+    def createTraceSpan(
         self,
         name: str,
         input: str,
         output: str,
         start_time: Optional[Nanoseconds] = None,
-        tags: Optional[Tags] = None,
     ) -> TraceSpan:
-        return TraceSpan(name, input, output, start_time, tags)
+        return TraceSpan(name, input, output, start_time)
