+++ python/packages/sdk/sdk/tests/test_generate_ids.py
@@ -1,4 +1,4 @@
-from ..generate_ids import generate_id
+from ..span.id import generate_id
 
 
 def test_generate_id():
