--- python/packages/sdk/sdk/trace_span.py
@@ -1,126 +1,0 @@
-from __future__ import annotations
-
-from time import time_ns
-from typing import List, Optional
-from contextvars import ContextVar
-
-from backports.cached_property import cached_property  # available in Python >=3.8
-from typing_extensions import Final, Self
-
-from .base import Nanoseconds, TraceId
-from .exceptions import (
-    ClosureOnClosedSpan,
-    FutureSpanStartTime,
-    InvalidType,
-    UnreachableTrace,
-)
-from .generate_ids import generate_id
-from .resource_name import get_resource_name
-from .tags import Tags
-
-
-__all__: Final[List[str]] = [
-    "TraceSpan",
-]
-
-
-TraceSpanContext = ContextVar[Optional["TraceSpan"]]
-
-
-class TraceSpan:
-    parentSpan: Self
-    name: str
-    startTime: Nanoseconds
-    endTime: Optional[Nanoseconds]
-    input: Optional[str]
-    output: Optional[str]
-    tags: Tags
-
-    def __init__(
-        self,
-        name: str,
-        input: Optional[str] = None,
-        output: Optional[str] = None,
-        start_time: Optional[Nanoseconds] = None,
-        tags: Optional[Tags] = None,
-    ):
-        self.name = get_resource_name(name)
-        self.input = input
-        self.output = output
-
-        self._set_start_time(start_time)
-        self._set_tags(tags)
-        self._set_parent_span()
-
-    def _set_parent_span(self):
-        global root_span
-
-        if root_span is None:
-            root_span = self
-            self.parentSpan = None
-
-        elif root_span.endTime is not None:
-            raise UnreachableTrace("Cannot initialize span: Trace is closed")
-
-        self.parentSpan = TraceSpan.resolveCurrentSpan()
-
-        while self.parentSpan.endTime:
-            self.parentSpan = self.parentSpan.parentSpan or root_span
-
-    def _set_tags(self, tags: Optional[Tags]):
-        self.tags = Tags()
-
-        if tags is not None:
-            self.tags.update(tags)
-
-    def _set_start_time(self, start_time: Optional[Nanoseconds]):
-        default_start = time_ns()
-
-        if start_time is not None and not isinstance(start_time, Nanoseconds):
-            raise InvalidType(f"`startTime` must be an integer.")
-
-        if start_time > default_start:
-            raise FutureSpanStartTime(
-                "Cannot initialize span: Start time cannot be set in the future"
-            )
-
-        self.startTime = start_time or default_start
-
-    @cached_property
-    def id(self) -> TraceId:
-        return generate_id()
-
-    @cached_property
-    def traceId(self) -> TraceId:
-        parent = self.parentSpan
-
-        return parent.traceId if parent else generate_id()
-
-    @property
-    def output(self) -> str:
-        return self._output
-
-    @output.setter
-    def output(self, value: str):
-        if not isinstance(value, str):
-            raise InvalidType(f"`output` must be a string.")
-
-        self._output = value
-
-    @staticmethod
-    def resolveCurrentSpan() -> Optional[TraceSpan]:
-        parent = rootContext.get()
-
-        return parent or root_span or None
-
-    def close(self, end_time: Optional[Nanoseconds]):
-        default: Nanoseconds = time_ns()
-
-        if self.endTime is not None:
-            raise ClosureOnClosedSpan("TraceSpan already closed.")
-
-        self.endTime = default if end_time is None else end_time
-
-
-rootContext: Final[TraceSpanContext] = ContextVar("rootContext", default=None)
-root_span: Optional["TraceSpan"] = None
